// packages/frontend/src/types/schemas/prisma-generated/inputTypeSchemas/SubscriptionCreateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SubscriptionStatusSchema } from './SubscriptionStatusSchema';
import { UserCreateNestedOneWithoutSubscriptionInputSchema } from './UserCreateNestedOneWithoutSubscriptionInputSchema';

export const SubscriptionCreateInputSchema: z.ZodType<Prisma.SubscriptionCreateInput> = z.object({
  id: z.string().uuid().optional(),
  stripeCustomerId: z.string().optional().nullable(),
  stripeSubscriptionId: z.string().optional().nullable(),
  status: z.lazy(() => SubscriptionStatusSchema),
  currentPeriodStart: z.coerce.date(),
  currentPeriodEnd: z.coerce.date(),
  cancelAtPeriodEnd: z.boolean().optional(),
  nextPlanPriceId: z.string().optional().nullable(),
  nextPlanStartDate: z.coerce.date().optional().nullable(),
  prorationAmount: z.number().optional().nullable(),
  createdAt: z.coerce.date().optional(),
  updatedAt: z.coerce.date().optional(),
  user: z.lazy(() => UserCreateNestedOneWithoutSubscriptionInputSchema)
}).strict();

export default SubscriptionCreateInputSchema;